[dagster-airlift] move tags to a neutral location
