+++ examples/experimental/dagster-airlift/dagster_airlift/constants.py
@@ -1,0 +1,3 @@
+MIGRATED_TAG = "airlift/task_migrated"
+DAG_ID_TAG = "airlift/dag_id"
+TASK_ID_TAG = "airlift/task_id"
