+++ examples/experimental/dagster-airlift/dagster_airlift_tests/unit_tests/core_tests/test_dag_defs.py
@@ -1,7 +1,7 @@
 from dagster import AssetKey, AssetSpec, Definitions, multi_asset
 from dagster._core.definitions.asset_key import CoercibleToAssetKey
+from dagster_airlift.constants import DAG_ID_TAG, TASK_ID_TAG
 from dagster_airlift.core import dag_defs, task_defs
-from dagster_airlift.core.utils import DAG_ID_TAG, TASK_ID_TAG
 
 
 def from_specs(*specs: AssetSpec) -> Definitions:
